Remove test code used for Travis
Previously, we have some tricks in codes to run it on Travis CI.

Since we've moved on to GitHub Actions, these codes are not
required anymore.
If we were to change CI back to Travis CI in the future,
we would use Docker image with Q4M and these codes would remain unnecessary.